Commiting what has been done for PLAN-4
Currently a user can enter their name, start date/end date, and distance for the itinerary. The information is stored  into the mongodb. Currently trying display what is entered so we can use it. Angularapp folder is the angular app on port 4200 and planit folder is the node.js/express files that implement the rest api