PLAN-4 PLAN-20 routing to another page and more filters
Basic start screen for filter done (ugly though). Added routing so now if you click on the 'go to test' button it will go to that component. The url can be  replaced with another url. Default url is the /distances url (to be changed later to something better). Added the budget and location filters