PLAN-1 You can now access basic filters on another page
When you enter info on the bottom part of the the page and hit the testing button it will redirect to a new page /test that has all the filters in the url as well. You can access these parts of the url to get that value, eg. you can get the distance from the url and save it and use it for something else